requirements.txt
# CodeVoice Requirements - Phase 1 MVP
# Install with: pip install -r requirements.txt

# ============================================
# AUDIO PROCESSING
# ============================================
pyaudio==0.2.14
soundfile==0.12.1

# ============================================
# SPEECH RECOGNITION (ASR)
# ============================================
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# ============================================
# INTENT CLASSIFICATION & NLP
# ============================================
transformers==4.36.0
scikit-learn==1.3.2

# ============================================
# VAD (Voice Activity Detection)
# ============================================
# Silero VAD is loaded via torch.hub

# ============================================
# ASYNC & CONCURRENCY
# ============================================
aiofiles==23.2.1

# ============================================
# CONFIGURATION & ENV
# ============================================
python-dotenv==1.0.0
pyyaml==6.0.1

# ============================================
# TESTING
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1

# ============================================
# UTILITIES & LOGGING
# ============================================
colorama==0.4.6
numpy==1.24.3
